Add The Ability To Set Connect And Request Timeouts
This commit adds the ability to set connect and request timeouts. Technically, connect timeouts could already be set before this by creating a jdk `HttpClient`, setting the connect timeout directly and passing it to the `JdkHttpClient.apply` function. However, because the JDK `HttpClient` sets the request timeout _per request_ on the `HttpRequest.Builder`, there was no way to set a request timeout.

Unfortunately, adding the ability to set these timeouts to the existing `apply` method would break binary compatibility. To avoid that, this commit also adds a `JdkHttpClientBuilder`, which allows for adding more settings in the future via this interface without breaking binary compatibility. Effectively the same as is done in the core `http4s` code, e.g, `BlazeClientBuilder`, etc.

In an effort to keep the API simple, the `JdkHttpClient.apply` method was deprecated. The intent is to have a single interface to set the settings and create a `Client`.

Some code was added to `JdkHttpClient` to handle converting between Scala and Java Duration types. Some which formerly were defined inside the `JdkHttpClient.apply` method were pulled out so they could be used by `JdkHttpClient.apply` and `JdkHttpClient.fromBuilder`.