Terminate CompletableFuture Values In All Exit States (#331)
* Terminate CompletableFuture Values In All Exit States

This commit adds a new function `fromCompletableFutureShift` and replaces the usage of `fromCompletionStage` with this function.

This function is similar to `fromCompletionStage`, except that it brackets on the effect, ensuring that in the case of cancellation or error the `CompletableFuture` is terminated with `.cancel` or `.completeExceptionally`.

Failing to do this, as `fromCompletionStage` did, can lead to a resource leak and extra CPU cycles when the effect which results from `fromCompletionStage` is terminated due to some external condition, e.g. `.timeout`.

To verify this, one needs only replace `fromCompletableFutureShift` in `CompletableFutureTerminationTest` with `org.http4s.internal.fromCompletionStage` and run the test.