Fix Resource Leak When The HTTPResponse Body Is Not Read
From the JDK docs for HttpResponse#BodyHandlers.ofPublisher,

https://docs.oracle.com/en/java/javase/15/docs/api/java.net.http/java/net/http/HttpResponse.BodyHandlers.html#ofPublisher()

> When the HttpResponse object is returned, the response headers will have been completely read, but the body may not have been fully received yet. The HttpResponse.body() method returns a Publisher<List<ByteBuffer>> from which the body response bytes can be obtained as they are received. The publisher can and must be subscribed to only once.

Of particular note is the final sentence, "The publisher can and must be subscribed to only once.".

There is a much longer discussion on the semantics of this in the comments in the code, but the short version is that we must always ensure that the `Flow.Publisher` returned on the `HttpResponse` when we use `BodyHandlers.ofPublisher` is subscribed to _exactly once_, whether or not anyone reads the body. Once we subscribe, the `Flow.Subscription` must be read or canceled.

Prior to this commit, we weren't doing that in cases where the body was not read, e.g. `Client#status`. For clarification, we _were_ (and still are) interrupting the body stream via a `.interruptWhen` call, but this doesn't subscribe to the `fs2.Stream` if no one has attempted to pull from it yet.

It is _really hard_ to write a unit test to demonstrate this behavior, but thankfully, it is _really easy_ to write some toy code which triggers it. To that end, see https://github.com/isomarcte/http4s-jdk-client-body-resource-leak , which is a repo which has a POC of this error. Running the `Main` method in there should trigger the issue in about 15 seconds. If you locally publish this commit and then re-run that with the this version, it will run forever using constant memory as intended.

For convenience, here are links to the relevant documentation or code,

* https://docs.oracle.com/en/java/javase/15/docs/api/java.net.http/java/net/http/HttpResponse.BodyHandlers.html#ofPublisher()
* https://github.com/typelevel/fs2/blob//reactive-streams/src/main/scala/fs2/interop/reactivestreams/StreamSubscriber.scala#L64
    * Note about how fs2 doesn't subscribe until someone starts pulling from the Stream.
* https://github.com/typelevel/fs2/blob//reactive-streams/src/main/scala/fs2/interop/reactivestreams/package.scala#L55
    * Shows how to make a fs2.Stream from a Publisher and attach an effect to the subscription event. This commit uses this to ensure that subscription always happens exactly once.